feat: move slog to utils folder
