+++ packages/substream/sink/events/subspaces-removed/handler.ts
@@ -2,7 +2,7 @@ import { Effect } from 'effect';
 
 import type { SubspaceRemoved } from './parser';
 import type { BlockEvent } from '~/sink/types';
-import { slog } from '~/sink/utils';
+import { slog } from '~/sink/utils/slog';
 
 export function handleSubspacesRemoved(_: SubspaceRemoved[], block: BlockEvent) {
   return Effect.gen(function* (_) {
